support cost calculator buttons
